requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
dnspython==2.1.0
email-validator==1.1.2
fake-useragent==0.1.11
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
idna==2.10
lxml==4.6.2
motor==2.3.0
parse==1.19.0
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
pyee==8.1.0
PyJWT==2.0.1
pymongo==3.11.3
pyppeteer==0.2.5
pyquery==1.4.3
python-decouple==3.4
python-dotenv==0.15.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
starlette==0.13.6
tqdm==4.56.0
urllib3==1.26.3
uvicorn==0.13.3
uvloop==0.14.0
w3lib==1.22.0
watchgod==0.6
websockets==8.1
